Update MetadataBundler to add support for module barrel files. Add tests.
The main limitation is that the `resolveModule` function assumes that the importName references
a TypeScript source file directly, e.g.

    './lib/foo' -> './lib/foo.ts'

This fails when using index.ts modules, e.g.

    './lib/foo' -> './lib/foo/index.ts'

To resolve this, the `resolveModulePath` function now probes for 'module.ts' and
'module/index.ts' patterns explicitly.

The value returned by `resolveModulePath` is relative to the root module.

The ModuleBundler class currently assumes that the root module is *NOT* of the
'module/index.ts' style - though adding a check to support this use case will
not be difficult.